Restore process main module
